requirements.txt
contentful==2.1.1
python-dotenv==1.0.0
Pillow==9.5.0
requests==2.31.0
